# Conflicts:
#	build.gradle

+++ src/main/java/com/kueennevercry/findex/dto/CursorPageResponseSyncJobDto.java
@@ -8,10 +8,11 @@ import lombok.RequiredArgsConstructor;
 @Getter
 @RequiredArgsConstructor
 public class CursorPageResponseSyncJobDto {
-private final List<SyncJobDto> content;
-private final String nextCursor;
-private final int nextIdAfter;
-private final int size;
-private final int totalElements;
-private final boolean hasNext;
+
+  private final List<SyncJobDto> content;
+  private final String nextCursor;
+  private final Long nextIdAfter;
+  private final int size;
+  private final Long totalElements;
+  private final boolean hasNext;
 }
